[ap3222][ag2922] Made recalculate scheduler values more efficient

+++ src/threads/fixed-point.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef int64_t fp_t;
+typedef int32_t fp_t;
 
 #define FP_Q 14
 #define FP_P 17
